refactor(login): move "login" folder to /shared
Refactor in folder structure to follow the guide

Closes #8
