+++ src/app/app.routing.ts
@@ -1,7 +1,7 @@
 import { Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
+import { LoginComponent } from './shared/login/login.component';
 import { RegisterComponent } from './shared/register/register.component';
 import { ExpenseListComponent } from './expenses/expense-list/expense-list.component';
 import { ExpenseComponent } from './expenses/expense/expense.component';
